pyproject.toml
[tool.poetry]
name = "erik"
version = "0.1.0"
description = ""
authors = ["teaishealthy <teaishealthy@protonmail.com>"]
license = "MIT"

[tool.pyright]
typeCheckingMode = "strict"
useLibraryCodeForTypes = true
reportMissingTypeStubs = false

[tool.poetry.dependencies]
python = "^3.8"
mcstatus = "^9.4.2"
"eludris.py" = {git = "https://github.com/teaishealthy/eludris.py"}
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
